01.02.2020
Added:
0. Web socket connection
1. Image fetch from URL
Need to add:
0. JSON parsing
1. Filtration
2. "No data" view
3. Save setting menu state

+++ AutoPosSyFAD/Model/SeparatorData.swift
@@ -12,6 +12,6 @@ struct SeparatorData {
     let date: String
     let type: Int
     let position: Int
-    let currentProbability: Double
+    let currentProbability: Int
     let img: String
 }
